requirements.txt
scikit-learn==1.1.3
pillow
numpy<2.0
scipy
pandas
scipy
